Moved http fakes into test base
