+++ tests/BVNetwork.404Handler.Tests/Base/Http/FakeHttpContext.cs
@@ -1,0 +1,14 @@
+﻿using System.Web;
+
+namespace BVNetwork.NotFound.Tests
+{
+    public partial class ErrorHandlerTests
+    {
+        public class FakeHttpContext : HttpContextBase
+        {
+            public override HttpRequestBase Request { get; } = new FakeHttpRequest();
+            public override HttpResponseBase Response { get; } = new FakeHttpResponse();
+            public override HttpServerUtilityBase Server { get; } = new FakeHttpServerUtility();
+        }
+    }
+}
