+++ tests/BVNetwork.404Handler.Tests/Base/Http/FakeHttpRequest.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Web;
+
+namespace BVNetwork.NotFound.Tests
+{
+    public partial class ErrorHandlerTests
+    {
+        public class FakeHttpRequest : HttpRequestBase
+        {
+            public override Uri Url { get; } = new Uri("http://example.com");
+        }
+    }
+}
