+++ tests/BVNetwork.404Handler.Tests/Base/Http/FakeHttpServerUtility.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Web;
+
+namespace BVNetwork.NotFound.Tests
+{
+    public partial class ErrorHandlerTests
+    {
+        public class FakeHttpServerUtility : HttpServerUtilityBase
+        {
+            public override Exception GetLastError()
+            {
+                return new Exception();
+            }
+        }
+    }
+}
