+++ tests/BVNetwork.404Handler.Tests/RequestHandlerTests.cs
@@ -1,0 +1,7 @@
+﻿namespace BVNetwork.NotFound.Tests
+{
+    public class RequestHandlerTests
+    {
+
+    }
+}
